+++ src/camera/view.c
@@ -38,7 +38,7 @@ t_matrix	*view_transform(t_tuple *from, t_tuple *to, t_tuple *up)
 	left = cross(forward, up);
 	true_up = cross(left, forward);
 	free(up);
-	orientation = new_matrix(4, (float [4][4]){
+	orientation = new_matrix(4, (double [4][4]){
 		{left->x, left->y, left->z, 0}, {true_up->x, true_up->y, true_up->z, 0},
 		{-forward->x, -forward->y, -forward->z, 0}, {0, 0, 0, 1},
 		});
