+++ src/tuple/tuple_creation.c
@@ -12,7 +12,7 @@
 
 #include <minirt.h>
 
-t_tuple	*new_tuple(float x, float y, float z, float w)
+t_tuple	*new_tuple(double x, double y, double z, double w)
 {
 	t_tuple	*tuple;
 
@@ -26,12 +26,12 @@ t_tuple	*new_tuple(float x, float y, float z, float w)
 	return (tuple);
 }
 
-t_tuple	*new_point(float x, float y, float z)
+t_tuple	*new_point(double x, double y, double z)
 {
 	return (new_tuple(x, y, z, 1.0));
 }
 
-t_tuple	*new_vector(float x, float y, float z)
+t_tuple	*new_vector(double x, double y, double z)
 {
 	return (new_tuple(x, y, z, 0.0));
 }
